fixup - add optional tenantID argument
